requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.6.3
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.1.2
frozenlist==1.4.0
future==0.18.3
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
inflect==7.2.1
lxml==4.9.3
more-itertools==10.3.0
multidict==6.0.4
pika==1.3.2
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.0.3
pydantic_core==2.3.0
pygramtic==0.2.0
pymongo==4.7.3
PyMySQL==1.1.1
pyotp==2.9.0
pyTelegramBotAPI==4.12.0
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2023.3
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rfc3986==1.5.0
robotframework==7.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.52
sseclient==0.0.27
telegram==0.0.1
tornado==6.1
typeguard==4.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzlocal==5.0.1
urllib3==2.0.3
websocket-client-py3==0.15.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.4.post2
